feat(tic-tac-toe): :sparkles: add game winning logic
