+++ src/components/TicTacToe/TicTacToe.tsx
@@ -4,11 +4,18 @@ import GameBoard from "../GameBoard/GameBoard";
 import GameLog from "../GameLog/GameLog";
 import Header from "../Header/Header";
 import ticTacToeStyles from "./TicTacToe.module.scss";
+import { WINNING_COMBINATIONS } from "../../constants/constants";
 
 interface gameTurns {
   square: { row: number; col: number };
   player: string;
 }
+
+const initialBoard = [
+  ["", "", ""],
+  ["", "", ""],
+  ["", "", ""],
+];
 
 const TicTacToe = () => {
   const [gameTurns, setGameTurns] = useState<gameTurns[]>([]);
@@ -23,6 +30,35 @@ const TicTacToe = () => {
   };
 
   const activePlayer = deriveActivePlayer(gameTurns);
+
+  const gameBoard = initialBoard;
+
+  for (const turn of gameTurns) {
+    const { square, player } = turn;
+    const { row, col } = square;
+
+    gameBoard[row][col] = player;
+  }
+
+  let winner;
+
+  for (const combination of WINNING_COMBINATIONS) {
+    const firstSquareSymbol =
+      gameBoard[combination[0].row][combination[0].column];
+    const secondSquareSymbol =
+      gameBoard[combination[1].row][combination[1].column];
+    const thirdSquareSymbol =
+      gameBoard[combination[2].row][combination[2].column];
+
+    if (firstSquareSymbol === "X" || firstSquareSymbol === "O") {
+      if (
+        firstSquareSymbol === secondSquareSymbol &&
+        secondSquareSymbol === thirdSquareSymbol
+      ) {
+        winner = firstSquareSymbol;
+      }
+    }
+  }
 
   const handleSelectSquare = (rowIndex: number, colIndex: number) => {
     setGameTurns((prevTurns) => {
@@ -43,10 +79,11 @@ const TicTacToe = () => {
   return (
     <section className={ticTacToeStyles.mainContainer}>
       <Header />
+      <p>You won {winner}</p>
       <GameBoard
         handleSelectSquare={handleSelectSquare}
         activePlayer={activePlayer}
-        turns={gameTurns}
+        board={gameBoard}
       />
       <GameLog turns={gameTurns} />
     </section>
