+++ src/components/TicTacToeBoard/TicTacToeBoard.tsx
@@ -2,28 +2,13 @@ import ticTacToeBoardStyles from "./TicTacToeBoard.module.scss";
 
 interface TicTacToeBoardProps {
   onSelectSquare: (row: number, col: number) => void;
-  turns: { square: { row: number; col: number }; player: string }[];
+  board: string[][];
 }
 
-const initialBoard = [
-  ["", "", ""],
-  ["", "", ""],
-  ["", "", ""],
-];
-
-const TicTacToeBoard = ({ onSelectSquare, turns }: TicTacToeBoardProps) => {
-  const gameBoard = initialBoard;
-
-  for (const turn of turns) {
-    const { square, player } = turn;
-    const { row, col } = square;
-
-    gameBoard[row][col] = player;
-  }
-
+const TicTacToeBoard = ({ onSelectSquare, board }: TicTacToeBoardProps) => {
   return (
     <ol className={ticTacToeBoardStyles.gameBoard}>
-      {gameBoard.map((row, index) => (
+      {board.map((row, index) => (
         <li key={index}>
           <ol className={ticTacToeBoardStyles.gameBoardRow}>
             {row.map((playerSymbol, colIndex) => (
